update ruby version
The Ruby version needed to be updated for Inspec to install properly.
I have fully converged everything but the Jenkins box with this version
of Ruby and have seen no warnings or issues

The Jenkins box was throwing errors for me before the Ruby swap so I am
confident my problems with Jenkins are unrelated. I will open a second PR
to address that issue.